Add files via upload
Visualisation for the computation. Two things that need to be done here are: 1) we need to make a call to the computational functions to get the path (Should be just one line of code) 2) Currently the latitude and longitudes are getting plotted as points, we can add lines between them - if someone could do it, it would be great since i was struggling to do that. (PS: Currently I have added an arbitrary list with float points, just to make sure our visualisation is working as it is.)